feat(App): consolidated WarehouseCLI with App and finalized the project
